Add ProjectController
Move project related routes out to their own controller
