+++ app/Http/Controllers/PagesController.php
@@ -4,8 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests;
 use App\Project;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Request;
 
 class PagesController extends Controller
 {
@@ -27,66 +25,6 @@ class PagesController extends Controller
 		return view('pages.projectfinder', ['projects' => Project::all()]);         //resources/views/pages/projectfinder.blade.php
 	}
 
-	public function project($id)
-	{
-        return view('pages.project', ['project' => Project::find($id)]);
-    }
-
-	public function create()
-	{
-		if (Auth::guest()) {
-			return redirect('/login');
-		}
-		return view('pages.create');
-	}
-
-	public function createProject()
-	{
-		if (Auth::guest()) {
-			return redirect('/login');
-		}
-		$title = Request::input('title');
-		$description = Request::input('description');
-		$body = Request::input('project-body');
-		$newEntry = Project::create([
-			'author' => Auth::user()->username,
-			'title'  => $title,
-			'description' => $description,
-			'body' => $body
-		]);
-
-		if (Request::hasFile('thumbnail')) {
-			$path = str_replace('/app', '', app_path());
-			$thumbnail = Request::file('thumbnail');
-			$thumbnail->move($path . '/public/images/projects',  'product' . $newEntry->id . '.jpg');
-		}
-
-		return redirect('/project/' . $newEntry->id);
-	}
-
-	public function deleteProject($id) {
-		if (Auth::guest()) {
-			return redirect('/login');
-		}
-
-		$project = Project::find($id);
-		if ($project == null) {
-			// TODO: Redirect to an error page
-			return redirect('/projectfinder');
-		}
-
-		if ($project->author == Auth::user()->username) {
-			$project->delete();
-			$imagePath = base_path() . '/public/images/projects/product' . $id . '.jpg';
-			if (file_exists($imagePath)) {
-				unlink($imagePath);
-			}
-		}
-
-		// TODO: Notice of success?
-		return redirect('/projectfinder');
-	}
- 
 	/*public function register()
 	{
 		return view('pages.register');             //resources/views/pages/register.blade.php
